A proposal of a new style of API together with a new style of front-end services.
The idea is to separate current models and APIs into logical blocks. All models, i.e. Mongoose schemas, should be placed in the models directory (/lib/models) and all the API methods in the API directory (/lib/api), both in their respective files. The example of such a structure is given with the Books schema and API. All the API  methods are simply REST services that can be called by anyone (given they are authorized where appropriate). With this approach, the authorization should be more flexible and easier to build in as it can be done on separate routes rather than their groups (although it will work on groups of routes too, of course).

This will, of course, affect the way this API is called from the Angular client. The one big APIservice will be split into smaller logical blocks and its services will be called in a new, more convenient way. An example can be seen in the Books service (/public/js/services/Books.js file) used from welcomeController.

All that should result in smaller files and a clearer, more sustainable and extensible code.